Move checkboxes out of Main Form
