+++ src/App.js
@@ -2,64 +2,66 @@ import React, { Component } from 'react';
 import _ from 'lodash';
 import { getFormValues } from 'redux-form';
 import { connect } from 'react-redux';
+import {
+  reduxForm,
+  formValueSelector,
+  change,
+} from 'redux-form';
 
 import './theme/App.css';
 import Chart from './components/Chart';
 import MainForm from './components/MainForm';
-import Return from './calculations/Return';
 import parseNumbersObject from './helpers/parseNumbersObject';
+import illnessEventAge from './calculations/illnessEventAge';
+import Checkboxes from './components/Checkboxes';
 
 class App extends Component {
-  ages() {
-    return _.range(this.props.inputs.currentAge, this.props.inputs.retirementAge + 1);
-  }
-
-  inputsValid() {
-    const {
-      currentAge,
-      retirementAge,
-    } = this.props.inputs;
-
-    if (!currentAge) return false;
-    if (!retirementAge) return false;
-    if (currentAge > retirementAge) return false;
-
-    return true;
-  }
-
-  result() {
-    if (!this.inputsValid()) return [];
-
-    return _.map(this.ages(), (age) => {
-      const portfoliosReturn = _.round(new Return(this.props.inputs,
-        'portfoliosFeesPercentage').calculate(age));
-      const mutualFundsReturn = _.round(new Return(this.props.inputs,
-        'mutualFundsFeesPercentage').calculate(age));
-
-      return ({
-        age,
-        portfoliosReturn,
-        mutualFundsReturn,
-        difference: portfoliosReturn - mutualFundsReturn,
-      });
-    });
-  }
-
   render() {
     return (
       <div style={styles.container}>
         <div style={styles.innerContainer}>
-          <MainForm />
-          <Chart data={this.result()} />
+          <MainForm {...this.props} />
+          <div style={styles.right}>
+            <Checkboxes />
+            <Chart {...this.props} />
+          </div>
         </div>
       </div>
     );
   }
 }
 
+const defaultCurrentAge = 30;
+const defaultRetirementAge = 65;
+
+const ReduxFormOnApp = reduxForm({
+  form: 'mainForm',
+  initialValues: {
+    currentAge: defaultCurrentAge,
+    retirementAge: defaultRetirementAge,
+    illnessEventAge: illnessEventAge(defaultCurrentAge, defaultRetirementAge),
+    primaryCiAmount: 69000,
+    primaryIllness: true,
+    secondaryCiAmount: 92000,
+    secondaryIllness: true,
+    rateOfReturnPercentage: 5,
+    portfoliosFeesPercentage: 0.68,
+    mutualFundsFeesPercentage: 2.3,
+    includePrimaryCiInsurance: true,
+    primaryCiCost: 33,
+    includeSecondaryCiInsurance: true,
+    secondaryCiCost: 47,
+    pacMonth: 550,
+    initialInvestment: 75000,
+  },
+})(App);
+
+const formSelector = formValueSelector('mainForm');
 export default connect(state => ({
-  inputs: parseNumbersObject(getFormValues('mainForm')(state)),
-}))(App);
+  currentAge: formSelector(state, 'currentAge'),
+  retirementAge: formSelector(state, 'retirementAge'),
+  illnessEventAge: formSelector(state, 'illnessEventAge'),
+}))(ReduxFormOnApp);
 
 const styles = {
   container: {
@@ -74,5 +76,10 @@ const styles = {
     justifyContent: 'center',
     maxWidth: '1000px',
     margin: '0 auto',
+  },
+  right: {
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
   },
 };
